RF: Wipe out old ConfigManager and replace with new one
