+++ datalad/cmd.py
@@ -413,9 +413,9 @@ def link_file_load(src, dst, dry_run=False):
 
 def get_runner(*args, **kwargs):
     # needs local import, because the ConfigManager itself needs the runner
-    from . import dlcfg
+    from . import cfg
     # TODO:  this is all crawl specific -- should be moved away
-    if dlcfg.obtain('datalad.crawl.dryrun', default=False):
+    if cfg.obtain('datalad.crawl.dryrun', default=False):
         kwargs = kwargs.copy()
         kwargs['protocol'] = DryRunProtocol()
     return Runner(*args, **kwargs)
