+++ datalad/interface/crawl.py
@@ -25,7 +25,7 @@ from datalad import utils
 from logging import getLogger
 lgr = getLogger('datalad.api.crawl')
 
-from .. import dlcfg
+from .. import cfg
 
 class Crawl(Interface):
     """Crawl online resource to create or update a dataset.
@@ -91,9 +91,9 @@ class Crawl(Interface):
             # TODO: centralize via _params_ handling
             if dry_run:
                 dryrun_optlabel = 'datalad.crawl.dryrun'
-                if dryrun_optlabel in dlcfg:
-                    dlcfg.unset(dryrun_optlabel, where='local', reload=False)
-                dlcfg.add(dryrun_optlabel, "True", where='local')
+                if dryrun_optlabel in cfg:
+                    cfg.unset(dryrun_optlabel, where='local', reload=False)
+                cfg.add(dryrun_optlabel, "True", where='local')
 
             if path is None:
