+++ datalad/metadata/__init__.py
@@ -25,7 +25,7 @@ from datalad.utils import swallow_logs
 from datalad.support.json_py import load as jsonload
 from datalad.dochelpers import exc_str
 from datalad.log import lgr
-from datalad import dlcfg
+from datalad import cfg
 
 
 # common format
@@ -314,7 +314,7 @@ def get_metadata(ds, guess_type=False, ignore_subdatasets=False,
 
 def _cached_load_document(url):
     from pyld.jsonld import load_document
-    cache_dir = opj(dlcfg.obtain('datalad.locations.cache'), 'schema')
+    cache_dir = opj(cfg.obtain('datalad.locations.cache'), 'schema')
     doc_fname = opj(
         cache_dir,
         '{}-{}'.format(
